Separate client&server js, add yaml w/ md parser
