+++ lib/data.js
@@ -1,47 +1,91 @@
-import { promises as fs } from 'fs';
 import path from 'path';
+import { loadSiteData } from './loader';
 
-import { ParserDatabase } from './parsers';
+export class Site {
 
-function basename(filepath) {
-    let almost = path.basename(filepath);
-    
-    // Check if still contains file extension, 
-    // if so remove it
-    let lastPeriod = almost.lastIndexOf(".");
-    if (lastPeriod != -1) {
-        almost = almost.substring(0, lastPeriod);
+    static DATA_PATH = path.join(process.cwd(), 'data/');
+    static DATA = null;
+
+    static async getData(...keys) {
+        if (Site.DATA === null) 
+            Site.DATA = await loadSiteData(Site.DATA_PATH);
+
+        let out = Site.DATA;
+        for (let key of keys) 
+            out = out[key];
+
+        return out;
     }
 
-    return almost;
 }
 
-export async function loadSiteData(root, data={}) {
+export class Conferences {
 
-    // Get the contents of the current path (as
-    // file objects)
-    const contents = await fs.readdir(
-        root, { withFileTypes: true }
-    );
-
-    // Parse contents or continue traversing
-    for (const content of contents) {
-        if (content.isFile()) {
-            const filepath = path.join(root, content.name);
-            
-            for (let parser of ParserDatabase) {
-                if (parser.canParse(filepath)) {
-                    data[basename(filepath)] = await parser.parseFile(filepath);
-                }
-            }
-        } 
-        else if (content.isDirectory()) {
-            let dirpath = path.join(root, content.name),
-                dirname = content.name;
-            
-            data[dirname] = await loadSiteData(dirpath);
-        }
+    static async getStuyConferences() {
+        return await Site.getData('conferences', 'stuy');
     }
-    
-    return data;
+
+    static async getStuyConference(name) {
+        return (await Conferences.getStuyConferences())[name];
+    }
+
+    static async getOtherConferences() {
+        return await Site.getData('conferences', 'other');
+    }
+
+    static DATA = null;
+
+    static async getConferences() {
+        if (Conferences.DATA === null) {
+            // Conferences.DATA = await Conferences.forceGetConferences();
+            return await Conferences.forceGetConferences();
+        }
+        return Conferences.DATA;
+    }
+
+    static async forceGetConferences() {
+        function* iterateObject(object) {
+            for (let key in object) {
+                yield [ key, object[key] ];
+            }
+        }
+        
+        function assignToObject(target, object, extraSauce) {
+            for (let [ key, value ] of iterateObject(object)) {
+                for (let [extraKey, extraValue] of iterateObject(extraSauce)) {
+                    value[extraKey] = extraValue;
+                }
+                target[key] = value;
+            }
+        }
+        
+        let conferences = {};
+
+        assignToObject(
+            conferences, 
+            await Conferences.getStuyConferences(), 
+            {type: 'stuy'}
+        );
+
+        assignToObject(
+            conferences, 
+            await Conferences.getOtherConferences(), 
+            {type: 'other'}
+        );
+
+        return conferences;
+    }
+
+    static async getConference(name) {
+        return (await Conferences.getConferences())[name];
+    }
+
+}
+
+export class News {
+
+    static async getPosts() {
+        return await Site.getData('news');
+    }
+
 }
