+++ lib/parsers.js
@@ -1,6 +1,7 @@
 import { promises as fs } from 'fs'; 
 import marked from 'marked';
 import yaml from 'js-yaml';
+import { readLines } from './util';
 
 class Parser {
     
@@ -25,24 +26,69 @@ class Parser {
 
 }
 
+function parseYml(text) {
+    let parsed = yaml.load(text),
+        what = typeof parsed;
+
+    if (what === 'string' || what === 'number')
+        return { content: parsed };
+    if (what === null)
+        return { content: null };
+
+    return parsed;
+}
+
+function parseMd(text) {
+    return { content: marked(text) };
+}
+
+function parseMdAndYml(text) {
+    let readingYml = false,
+        hasYml = false,
+        firstLine = true;
+
+    let lines = readLines(text),
+        mdBuffer = '',
+        ymlBuffer = '';
+
+    for (let line of lines) {
+        if (firstLine) {
+            firstLine = false;
+            readingYml = (line === '---\n');
+            hasYml = true;
+
+            if (readingYml)
+                continue;
+        }
+        
+        if (readingYml && line === '---\n') {
+            readingYml = false;
+            continue;
+        }
+
+        if (readingYml) 
+            ymlBuffer += line;
+        else    
+            mdBuffer += line;
+    }
+
+    let parsed = {
+        content: parseMd(mdBuffer).content
+    };
+
+    if (hasYml && ymlBuffer != '') 
+        parsed.metadata = parseYml(ymlBuffer)
+
+    return parsed;
+}
+
 export const ParserDatabase = [
     new Parser('.json', JSON.parse), 
 
-    new Parser('.yml', text => {
-        let parsed = yaml.load(text),
-            what = typeof parsed;
+    new Parser('.yml', parseYml),
 
-        if (what === 'string' || what === 'number')
-            return { content: parsed };
-        if (what === null)
-            return { content: null };
-
-        return parsed;
-    }),
-    
-    new Parser('.md', markdown => {
-        return { content: marked(markdown) };
-    }),
+    new Parser('.md', parseMd),
+    new Parser('.ymd', parseMdAndYml),
     
     new Parser('.html', html => {
         return { content: html };
