+++ lib/util.js
@@ -1,88 +1,31 @@
-import path from 'path';
-import { loadSiteData } from './data';
-
 export function isHttpLink(link) {
     return link !== undefined && 
             link.startsWith("http");
 }
 
-export class Site {
+export function* readLines(fileContent) {
+    let line = '';
+    for (let i = 0; i < fileContent.length; ++i) {
+        let char = fileContent.charAt(i);
+        line += char;
 
-    static DATA_PATH = path.join(process.cwd(), 'data/');
-    static DATA = null;
-
-    static async getData(...keys) {
-        if (Site.DATA === null) 
-            Site.DATA = await loadSiteData(Site.DATA_PATH);
-
-        let out = Site.DATA;
-        for (let key of keys) 
-            out = out[key];
-
-        return out;
+        if (char == '\n') {
+            yield line;
+            line = '';
+        }
     }
-
 }
 
-export class Conferences {
+// export function* toIterator(object) {
+//     for (let key in object) {
+//         yield [ key, object[key] ];    
+//     }
+// }
 
-    static async getStuyConferences() {
-        return await Site.getData('conferences', 'stuy');
-    }
-
-    static async getStuyConference(name) {
-        return (await Conferences.getStuyConferences())[name];
-    }
-
-    static async getOtherConferences() {
-        return await Site.getData('conferences', 'other');
-    }
-
-    static DATA = null;
-
-    static async getConferences() {
-        if (Conferences.DATA === null) {
-            // Conferences.DATA = await Conferences.forceGetConferences();
-            return await Conferences.forceGetConferences();
-        }
-        return Conferences.DATA;
-    }
-
-    static async forceGetConferences() {
-        function* iterateObject(object) {
-            for (let key in object) {
-                yield [ key, object[key] ];
-            }
-        }
-        
-        function assignToObject(target, object, extraSauce) {
-            for (let [ key, value ] of iterateObject(object)) {
-                for (let [extraKey, extraValue] of iterateObject(extraSauce)) {
-                    value[extraKey] = extraValue;
-                }
-                target[key] = value;
-            }
-        }
-        
-        let conferences = {};
-
-        assignToObject(
-            conferences, 
-            await Conferences.getStuyConferences(), 
-            {type: 'stuy'}
-        );
-
-        assignToObject(
-            conferences, 
-            await Conferences.getOtherConferences(), 
-            {type: 'other'}
-        );
-
-        return conferences;
-    }
-
-    static async getConference(name) {
-        return (await Conferences.getConferences())[name];
-    }
-
-}
+// export function map(iterator, callback) {
+//     let mappedItems = [], i = 0;
+//     for (let item of iterator) {
+//         mappedItems.push(callback(item, ++i));
+//     }
+//     return mappedItems;
+// }
